28639: Clean up generate_exception_blob
