+++ src/hotspot/cpu/loongarch/sharedRuntime_loongarch_64.cpp
@@ -2931,4 +2931,138 @@ RuntimeStub* SharedRuntime::generate_resolve_blob(address destination, const cha
   return tmp;
 }
 
+#ifdef COMPILER2
+//-------------- generate_exception_blob -----------
+// creates exception blob at the end
+// Using exception blob, this code is jumped from a compiled method.
+// (see emit_exception_handler in loongarch.ad file)
+//
+// Given an exception pc at a call we call into the runtime for the
+// handler in this method. This handler might merely restore state
+// (i.e. callee save registers) unwind the frame and jump to the
+// exception handler for the nmethod if there is no Java level handler
+// for the nmethod.
+//
+// This code is entered with a jump, and left with a jump.
+//
+// Arguments:
+//   A0: exception oop
+//   A1: exception pc
+//
+// Results:
+//   A0: exception oop
+//   A1: exception pc in caller
+//   destination: exception handler of caller
+//
+// Note: the exception pc MUST be at a call (precise debug information)
+//
+//  [stubGenerator_loongarch_64.cpp] generate_forward_exception()
+//      |- A0, A1 are created
+//      |- T4 <= SharedRuntime::exception_handler_for_return_address
+//      `- jr T4
+//           `- the caller's exception_handler
+//                 `- jr OptoRuntime::exception_blob
+//                        `- here
+//
+void OptoRuntime::generate_exception_blob() {
+  enum frame_layout {
+    fp_off, fp_off2,
+    return_off, return_off2,
+    framesize
+  };
+  assert(framesize % 4 == 0, "sp not 16-byte aligned");
+
+  // Allocate space for the code
+  ResourceMark rm;
+  // Setup code generation tools
+  CodeBuffer buffer("exception_blob", 2048, 1024);
+  MacroAssembler* masm = new MacroAssembler(&buffer);
+
+  address start = __ pc();
+
+  // Exception pc is 'return address' for stack walker
+  __ push2(A1 /* return address */, FP);
+
+  // there are no callee save registers and we don't expect an
+  // arg reg save area
+#ifndef PRODUCT
+  assert(frame::arg_reg_save_area_bytes == 0, "not expecting frame reg save area");
+#endif
+  // Store exception in Thread object. We cannot pass any arguments to the
+  // handle_exception call, since we do not want to make any assumption
+  // about the size of the frame where the exception happened in.
+  __ st_d(A0, Address(TREG, JavaThread::exception_oop_offset()));
+  __ st_d(A1, Address(TREG, JavaThread::exception_pc_offset()));
+
+  // This call does all the hard work.  It checks if an exception handler
+  // exists in the method.
+  // If so, it returns the handler address.
+  // If not, it prepares for stack-unwinding, restoring the callee-save
+  // registers of the frame being removed.
+  //
+  // address OptoRuntime::handle_exception_C(JavaThread* thread)
+  //
+  Label L;
+  address the_pc = __ pc();
+  __ bind(L);
+  __ set_last_Java_frame(TREG, SP, NOREG, L);
+
+  assert(StackAlignmentInBytes == 16, "must be");
+  __ bstrins_d(SP, R0, 3, 0);   // Fix stack alignment as required by ABI
+
+  __ move(A0, TREG);
+  __ call(CAST_FROM_FN_PTR(address, OptoRuntime::handle_exception_C),
+          relocInfo::runtime_call_type);
+
+  // handle_exception_C is a special VM call which does not require an explicit
+  // instruction sync afterwards.
+
+  // Set an oopmap for the call site.  This oopmap will only be used if we
+  // are unwinding the stack.  Hence, all locations will be dead.
+  // Callee-saved registers will be the same as the frame above (i.e.,
+  // handle_exception_stub), since they were restored when we got the
+  // exception.
+
+  OopMapSet *oop_maps = new OopMapSet();
+
+  oop_maps->add_gc_map(the_pc - start, new OopMap(framesize, 0));
+
+  __ reset_last_Java_frame(TREG, false);
+
+  // Restore callee-saved registers
+
+  // FP is an implicitly saved callee saved register (i.e. the calling
+  // convention will save restore it in prolog/epilog) Other than that
+  // there are no callee save registers now that adapter frames are gone.
+  // and we dont' expect an arg reg save area
+  __ pop2(RA, FP);
+
+  const Register exception_handler = T4;
+
+  // We have a handler in A0, (could be deopt blob)
+  __ move(exception_handler, A0);
+
+  // Get the exception
+  __ ld_d(A0, Address(TREG, JavaThread::exception_oop_offset()));
+  // Get the exception pc in case we are deoptimized
+  __ ld_d(A1, Address(TREG, JavaThread::exception_pc_offset()));
+#ifdef ASSERT
+  __ st_d(R0, Address(TREG, JavaThread::exception_handler_pc_offset()));
+  __ st_d(R0, Address(TREG, JavaThread::exception_pc_offset()));
+#endif
+  // Clear the exception oop so GC no longer processes it as a root.
+  __ st_d(R0, Address(TREG, JavaThread::exception_oop_offset()));
+
+  // A0: exception oop
+  // A1: exception pc
+  __ jr(exception_handler);
+
+  // make sure all code is generated
+  masm->flush();
+
+  // Set exception blob
+  _exception_blob = ExceptionBlob::create(&buffer, oop_maps, framesize >> 1);
+}
+#endif // COMPILER2
+
 extern "C" int SpinPause() {return 0;}
